pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oncall-agent"
version = "0.1.0"
description = "AI-powered Data Engineering OnCall Agent"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.1.0",
    "duckdb>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0" 
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
